feat(core): run HTML checks in headless Chrome
Replaces NightmareJS (driver for Electron) by Puppeteer (driver for headless Chrome).
Puppeteer is a Node API for headless Chrome, developped and officially supported
by the Chrome DevTools team.

See also: https://github.com/GoogleChrome/puppeteer

Closes #23
