+++ package.json
@@ -59,8 +59,8 @@
     "marko": "^4.4.26",
     "meow": "^3.7.0",
     "multer": "^1.3.0",
-    "nightmare": "^2.10.0",
     "path": "^0.12.7",
+    "puppeteer": "^0.12.0",
     "shortid": "^2.2.8",
     "snyk": "^1.42.7",
     "tmp": "^0.0.33",
@@ -84,9 +84,6 @@
     "standard-version": "^4.2.0",
     "uglify-js": "^3.0.8",
     "watch": "^1.0.2"
-  },
-  "resolutions": {
-    "nightmare/electron": "1.7.8"
   },
   "scripts": {
     "start": "node src/cli/cli.js",
